Desugar before infer ops
